helmet and refactoring
