+++ README.md
@@ -1,12 +1,19 @@
 # news-explorer-api
 
+# Overview
+* Intro
+* Features
+* Technologies
+* Deployment
+
+## Intro
 This is the backend for my final project as a Practicum student in the Web Development Program. This project is an article collection app, where users can search for news articles and save them to their profiles.
 
 ## Features
 - Routes for signup and signin, with user JWT authentication.
 - Routes for saving articles to profile, saved articles, and deleting saved articles. 
-- Mongo DB for saving users and articles.
-- Implementing logging 
+- Mongo DB for user profiles and articles saved.
+- Implemented logging 
 - Centralized error handling
 
 
@@ -16,3 +23,5 @@ This is the backend for my final project as a Practicum student in the Web Devel
 * Mongo
 
 ## Deployment
+jayme-news.
+api.jayme-news.
