+++ config/errors/errorMessages.js
@@ -1,0 +1,14 @@
+const errorMessages = {
+  authRequired = 'Authorization Required',
+  couldNotCreateArticle = 'Article validation failed',
+  notUsersArticle = 'Article does not belong to user',
+  invalidArticle = 'Invalid article',
+  invalidUser = 'User ID not found',
+  incorrectLogin = 'Incorrect password or email',
+
+
+}
+
+module.exports = errorMessages;
+
+// import error messages & error throw new error(errorMessage.notFoundError);
