+++ middleware/auth.js
@@ -1,12 +1,13 @@
 const jwt = require('jsonwebtoken');
 const { NODE_ENV, JWT_SECRET } = process.env;
-const UnauthorizationErr = require('../errors/unauth-err');
+const UnauthorizationErr = require('../config/errors/unauth-err');
+const errorMessages = require('../config/errors/errorMessages');
 
 module.exports = (req, res, next) => {
   const { authorization } = req.headers;
 
   if(!authorization || !authorization.startsWith('Bearer ')) {
-    throw new UnauthorizationErr('Authorization Required');
+    throw new UnauthorizationErr(errorMessages.authRequired);
   }
 
   const token = authorization.replace('Bearer ', '');
@@ -14,7 +15,7 @@ module.exports = (req, res, next) => {
   try {
     payload = jwt.verify(token, NODE_ENV === 'production' ? JWT_SECRET : 'dev-secret');
   } catch (err) {
-    throw new UnauthorizationErr('Authorization Required');
+    throw new UnauthorizationErr(errorMessages.authRequired);
   }
   req.user = payload;
   next();
